requirements.txt
# python 3.10

PyYAML~=5.4.1
torch~=2.0.1
gymnasium~=0.29.1
numpy~=1.25.2
neptune~=1.6.2